refactor(logger): move logger to lib directory
